feat: move methods for comments to independent file and add shutdown method
